# openapi/favouritezoo-api.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/favouritezoo",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/animals": {
      "get": {
        "tags": [
          "animal-controller"
        ],
        "operationId": "listAllAnimals",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimalResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "animal-controller"
        ],
        "operationId": "createAnimal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimalDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              }
            }
          }
        }
      }
    },
    "/animals/{animalId}": {
      "get": {
        "tags": [
          "animal-controller"
        ],
        "operationId": "findAnimal",
        "parameters": [
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "animal-controller"
        ],
        "operationId": "updateAnimal",
        "parameters": [
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimalDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "202": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "animal-controller"
        ],
        "operationId": "deleteAnimal",
        "parameters": [
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{roomId}/animals/{animalId}": {
      "patch": {
        "tags": [
          "animal-room-controller"
        ],
        "operationId": "placeAnimalIntoRoom",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/animals/{animalId}": {
      "patch": {
        "tags": [
          "animal-room-controller"
        ],
        "operationId": "removeRoomForAnimal",
        "parameters": [
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{roomId}/animals/{animalId}/{newRoomId}": {
      "patch": {
        "tags": [
          "animal-room-controller"
        ],
        "operationId": "moveAnimalBetweenRooms",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "newRoomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{roomId}/animals": {
      "get": {
        "tags": [
          "animal-room-controller"
        ],
        "operationId": "listAllAnimalsInRoom",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimalResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/happyAnimals": {
      "get": {
        "tags": [
          "animal-room-controller"
        ],
        "operationId": "listHappyAnimals",
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/favourite/animals/{animalId}": {
      "get": {
        "tags": [
          "favourite-controller"
        ],
        "operationId": "listOfFavoriteRooms",
        "parameters": [
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/favourite/rooms/{roomId}/animals/{animalId}": {
      "get": {
        "tags": [
          "favourite-controller"
        ],
        "operationId": "findAssignedFavouriteRoomForAnimal",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Favourite"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "favourite-controller"
        ],
        "operationId": "assignFavouriteRoomForAnimal",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Favourite"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "favourite-controller"
        ],
        "operationId": "unAssignFavouriteRoomForAnimal",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          },
          "204": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "GenericResponseDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/Sort"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "Sort": {
        "type": "object",
        "properties": {
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "AnimalResponseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "added": {
            "type": "string",
            "format": "date-time"
          },
          "located": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AnimalDto": {
        "required": [
          "title",
          "type"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "preference": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Animal": {
        "required": [
          "preference",
          "title",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "preference": {
            "type": "integer",
            "format": "int64"
          },
          "added": {
            "type": "string",
            "format": "date-time"
          },
          "located": {
            "type": "string",
            "format": "date-time"
          },
          "room": {
            "$ref": "#/components/schemas/Room"
          }
        }
      },
      "Room": {
        "required": [
          "size",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "animals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Animal"
            }
          }
        }
      },
      "Favourite": {
        "required": [
          "animal",
          "room"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "room": {
            "$ref": "#/components/schemas/Room"
          },
          "animal": {
            "$ref": "#/components/schemas/Animal"
          }
        }
      },
...
    }
  }
}