# src/main/java/com/jayaprabahar/favouritezoo/model/Animal.java
/**
 * 
 */
package com.jayaprabahar.favouritezoo.model;

import java.time.LocalDateTime;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.OneToOne;
import javax.validation.constraints.NotNull;

import org.hibernate.annotations.CreationTimestamp;
import org.hibernate.annotations.UpdateTimestamp;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * <p> Project : favouritezoo </p>
 * <p> Title : Animal.java </p>
 * <p> Description: Animal Entity </p>
 * <p> Created: Nov 10, 2020 </p>
 * 
 * @since 1.0.0
 * @version 1.0.0
 * @author <a href="mailto:jpofficial@gmail.com">Jayaprabahar</a>
 *
 */
@Entity
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class Animal {

	@Id
	@GeneratedValue(strategy = GenerationType.IDENTITY)
	@Column(name = "animal_id")
	private long id;

	@NotNull
	private String title;

	@NotNull
	private String type;

	@NotNull
	private long preference;

	@CreationTimestamp
	private LocalDateTime added;

	@UpdateTimestamp
	private LocalDateTime located;

	@OneToOne(targetEntity = Room.class)
	private Room room;

}